Updated transform system

+++ nectar_render/csrc/core/include/core/constants.h
@@ -4,10 +4,10 @@
 // NUMERICAL CONSTANTS
 // ############################################################################
 
-const float EPS  = (float)1e-3f;
+const float EPS  = 1e-3f;
 const float INF  = __FLT_MAX__;
-const float FMIN = (float)1e-4f;
-const float FMAX = (float)1e32f;
+const float FMIN = 1e-4f;
+const float FMAX = 1e32f;
 const float PI   = 3.1415926535897932385f;
 const float PI2  = PI * 2.0f;
